--- src/client/src/client/mod.rs
@@ -1,6 +1,0 @@
-#[allow(clippy::module_inception)]
-mod client;
-mod cluster;
-
-pub use client::*;
-pub use cluster::*;
