--- src/client/src/spu/mod.rs
@@ -1,3 +1,0 @@
-mod spu_pool;
-
-pub use spu_pool::*;
